requirements.txt
pydantic==1.10.13
exceptiongroup
feedparser
requests
python-dotenv
beautifulsoup4
readability-lxml
openai